# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tags": {
      "get": {
        "operationId": "TagController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TagList"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "一個標籤組的標籤；要進得了那個標籤組（例：file 組要能進檔案管理器）",
        "tags": ["tags"]
      },
      "post": {
        "operationId": "TagController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Tag"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["tags"]
      }
    },
    "/tags/{id}": {
      "patch": {
        "operationId": "TagController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Tag"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "改名、改色；標籤組不能改",
        "tags": ["tags"]
      },
      "delete": {
        "operationId": "TagController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "刪除；所有資源上的這個標籤一併移除，不進回收桶",
        "tags": ["tags"]
      }
    },
    "/tags/assignments/{resourceType}/{resourceId}": {
      "put": {
        "operationId": "TagController_replace",
        "parameters": [
          {
            "name": "resourceType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceResourceTagsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ResourceTags"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "整批取代一個資源的標籤；能不能改跟著那個資源的編輯權限",
        "tags": ["tags"]
      }
    },
...
  },
  "info": {
    "title": "B2B System API",
    "description": "RBAC 骨架 API",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TagSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "enum": ["neutral", "brand", "success", "warning", "danger"]
          }
        },
        "required": ["id", "name", "color"],
        "additionalProperties": false
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "scope": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "enum": ["neutral", "brand", "success", "warning", "danger"]
          },
          "version": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": ["id", "scope", "name", "color", "version", "createdAt", "updatedAt"],
        "additionalProperties": false
      },
      "TagList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          }
        },
        "required": ["items"],
        "additionalProperties": false
      },
      "CreateTagRequest": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "maxLength": 50,
            "pattern": "^[a-z][A-Za-z0-9]*$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "color": {
            "default": "neutral",
            "type": "string",
            "enum": ["neutral", "brand", "success", "warning", "danger"]
          }
        },
        "required": ["scope", "name", "color"],
        "additionalProperties": false
      },
      "UpdateTagRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "enum": ["neutral", "brand", "success", "warning", "danger"]
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        },
        "required": ["version"],
        "additionalProperties": false
      },
      "ReplaceResourceTagsRequest": {
        "type": "object",
        "properties": {
          "tagIds": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": ["tagIds"],
        "additionalProperties": false
      },
      "ResourceTags": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagSummary"
            }
          }
        },
        "required": ["tags"],
        "additionalProperties": false
      },
...
    }
  }
}